+++ test/examples/02.cloning.spec.ts
@@ -22,14 +22,14 @@ describe('cloning', () => {
     it('calls fetch')
 
     it.clone('successful response', 'some result', it => {
-      it('puts successAction', ({ value: { payload } }) => {
+      it('puts successAction', ({ payload }) => {
         payload.action.should.deep.equal(successAction('some result'))
       })
     })
 
     const error = new Error()
     it.clone('error response', error, it => {
-      it('puts errorAction', ({ value: { payload } }) => {
+      it('puts errorAction', ({ payload }) => {
         payload.action.should.deep.equal(errorAction(error))
       })
     })